+++ flow_control/__init__.py
@@ -2,4 +2,4 @@
 Init for main Flow-Control XBlock
 """
 from .flow import FlowCheckPointXblock
-__version__ = '1.0.1'
+__version__ = '2.0.0'