+++ flow_control/edxapp_wrapper/backends/score_s_v1.py
@@ -1,0 +1,15 @@
+"""
+Score module definitions for Open edX Sumac release.
+"""
+# pylint: disable=import-error
+from lms.djangoapps.courseware.model_data import ScoresClient
+
+
+def get_score_module(*args, **kwargs):
+    """
+    Get ScoresClient model.
+
+    Returns:
+        ScoresClient: ScoresClient object.
+    """
+    return ScoresClient(*args, **kwargs)