+++ flow_control/settings/test.py
@@ -1,0 +1,56 @@
+"""
+Django settings for flow_control project.
+For more information on this file, see
+https://docs.djangoproject.com/en/2.2/topics/settings/
+For the full list of settings and their values, see
+https://docs.djangoproject.com/en/2.2/ref/settings/
+"""
+
+DATABASES = {
+    "default": {
+        "ENGINE": "django.db.backends.sqlite3",
+        "NAME": "default.db",
+        "USER": "",
+        "PASSWORD": "",
+        "HOST": "",
+        "PORT": "",
+    },
+    "read_replica": {
+        "ENGINE": "django.db.backends.sqlite3",
+        "NAME": "read_replica.db",
+        "USER": "",
+        "PASSWORD": "",
+        "HOST": "",
+        "PORT": "",
+    },
+}
+
+SECRET_KEY = "not-so-secret-key"
+
+# Application definition
+
+INSTALLED_APPS = (
+    'flow_control',
+)
+
+# Internationalization
+# https://docs.djangoproject.com/en/2.2/topics/i18n/
+
+LANGUAGE_CODE = 'en-us'
+
+TIME_ZONE = 'UTC'
+
+USE_I18N = True
+
+USE_L10N = True
+
+USE_TZ = True
+
+
+# Static files (CSS, JavaScript, Images)
+# https://docs.djangoproject.com/en/2.2/howto/static-files/
+
+STATIC_URL = '/static/'
+
+# Plugin settings
+FLOW_CONTROL_SCORE_MODULE_BACKEND = 'flow_control.edxapp_wrapper.backends.score_s_v1'