+++ CHANGELOG.rst
@@ -14,6 +14,36 @@ Unreleased
 
 *
 
+[2.0.0] - 2025-01-23
+**********************************************
+
+Added
+=====
+
+* Support for Python 3.11 and Django 4.2
+* Add  github workflow
+
+Removed
+_______
+
+* **BREAKING CHANGE**: Dropped support for Python 3.5
+* Drop CircleCI support
+* Remove support to "redirect to another unit in the same subsection" action
+
+Changed
+=======
+
+* Fix UI issues
+
+[1.0.1] - 2020-10-14
+~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+
+Fixed
+_____
+
+* Python3 and Juniper issues
+* Fix applyFlowControl in injection to applied correctly the actions of flow-control 
+
 
 [1.0.0] - 2020-06-18
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~